# apps/tourist/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vualiku XP Distribution API",
        "version": "1.0.0",
        "description": "API for external B2B partners to list activities, check availability, and create reservations."
    },
    "servers": [
        {
            "url": "https://vualikuxp.com/api/v1",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:3000/api/v1",
            "description": "Local Development Server"
        }
    ],
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-Key"
            }
        }
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ],
    "paths": {
        "/activities": {
            "get": {
                "summary": "List Activities",
                "description": "Returns a catalog of all available activities.",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "count": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "operatorId": {
                                                        "type": "string"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "price": {
                                                        "type": "number"
                                                    },
                                                    "currency": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/availability/{date}": {
            "get": {
                "summary": "Check Availability",
                "description": "Returns availability for all activities on a specific date.",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "description": "Target date in YYYY-MM-DD format"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/reserve": {
            "post": {
                "summary": "Create Reservation",
                "description": "Creates a confirmed booking for a B2B partner.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "firstName",
                                    "lastName",
                                    "email",
                                    "items"
                                ],
                                "properties": {
                                    "firstName": {
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "origin": {
                                        "type": "string"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "required": [
                                                "eventName",
                                                "operatorId",
                                                "pax",
                                                "date"
                                            ],
                                            "properties": {
                                                "eventName": {
                                                    "type": "string"
                                                },
                                                "operatorId": {
                                                    "type": "string"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "timeSlot": {
                                                    "type": "string"
                                                },
                                                "pax": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reservation created"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        }
    }
}